move node-wot-helpers to node-wot core
