--- packages/node-wot-helpers/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "node-wot-helpers",
-  "version": "0.3.0-SNAPSHOT.2",
-  "files": [
-    "dist/"
-  ],
-  "description": "Helper functions used throughout the stack",
-  "main": "dist/helpers.js",
-  "typings": "dist/helpers.d.ts",
-  "scripts": {
-    "test": "mocha --compilers ts:ts-node/register",
-    "build": "tsc",
-    "codestyle" : "standard --pretty"
-  },
-  "author": "the thingweb community",
-  "license": "W3C Software License",
-  "devDependencies": {
-    "@types/chai": "4.0.4",
-    "@types/node": "8.0.28",
-    "chai": "4.1.2",
-    "mocha": "3.5.3",
-    "mocha-typescript": "1.1.8",
-    "ts-node": "3.3.0",
-    "typescript": "2.5.2",
-    "typescript-standard": "0.3.30"
-  },
-  "dependencies": {
-  }
-}
