--- packages/node-wot-helpers/tsconfig.json
@@ -1,18 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "es5",
-		"lib" : ["ES2015","ScriptHost"],
-		"module": "commonjs",
-		"outDir": "dist",
-		"sourceMap": true,
-		"noImplicitAny": true,
-		"removeComments": true,
-		"noLib": false,
-		"preserveConstEnums": true,
-		"experimentalDecorators": true,
-		"declaration": true
-	},
-	"include": [
-		"src/**/*"
-	]
-}
