+++ packages/node-wot/src/helpers.ts
