+++ packages/node-wot/src/index.ts
@@ -43,3 +43,6 @@ export * from 'node-wot-td-tools'
 
 export {HttpServer} from 'node-wot-protocol-http'
 export {HttpClient} from 'node-wot-protocol-http'
+
+// Helper Implementations
+export * from './helpers'
