+++ packages/node-wot/src/servient.ts
@@ -23,7 +23,7 @@ import WoTImpl from "./wot-impl";
 import {ProtocolClientFactory, ProtocolServer, ResourceListener, ProtocolClient} from "node-wot-protocols"
 import {ThingDescription} from "node-wot-td-tools";
 import * as TD from "node-wot-td-tools";
-import * as Helpers from "node-wot-helpers";
+import * as Helpers from "./helpers";
 import { default as ContentSerdes, ContentCodec } from "node-wot-content-serdes"
 import * as vm from 'vm'
