+++ packages/node-wot/test/HelpersTest.ts
