+++ packages/node-wot/tsconfig.json
@@ -1,7 +1,7 @@
 {
 	"compilerOptions": {
 		"target": "es5",
-		"lib" : ["es2015","dom"],
+		"lib" : ["es2015","dom","ScriptHost"],
 		"module": "commonjs",
 		"outDir": "dist",
 		"sourceMap": true,
